Check if ticket has owner before calling it.

+++ src/main/java/ch/wisv/areafiftylan/products/service/TicketServiceImpl.java
@@ -202,11 +202,13 @@ public class TicketServiceImpl implements TicketService {
         }
 
         // Remove User from his teams when transferring his last ticket away.
-        String ownerUsername = t.getOwner().getUsername();
-        if (findValidTicketsByOwnerUsername(ownerUsername).size() == 1) {
-            teamService.getTeamsByUsername(ownerUsername).stream().
-                    map(Team::getId).
-                    forEach((teamId) -> teamService.removeMember(teamId, ownerUsername));
+        if (t.getOwner() != null) {
+            String ownerUsername = t.getOwner().getUsername();
+            if (findValidTicketsByOwnerUsername(ownerUsername).size() == 1) {
+                teamService.getTeamsByUsername(ownerUsername).stream().
+                        map(Team::getId).
+                        forEach((teamId) -> teamService.removeMember(teamId, ownerUsername));
+            }
         }
 
         User newOwner = ttt.getUser();
